refactor(modeling): introduce EventIdCapable and SnapshotTimeCapable interfaces
- Add EventIdCapable interface for aggregates with event ID information
- Add SnapshotTimeCapable interface for aggregates with snapshot time information
- Update MaterializedSnapshot to implement new interfaces
- Create SlimMaterializedSnapshot for simplified snapshot representation
- Update Snapshot interface to implement SnapshotTimeCapable
