+++ wow-api/src/main/kotlin/me/ahoo/wow/api/query/MaterializedSnapshot.kt
@@ -15,12 +15,14 @@ package me.ahoo.wow.api.query
 
 import me.ahoo.wow.api.Version
 import me.ahoo.wow.api.modeling.DeletedCapable
+import me.ahoo.wow.api.modeling.EventIdCapable
 import me.ahoo.wow.api.modeling.EventTimeCapable
 import me.ahoo.wow.api.modeling.FirstEventTimeCapable
 import me.ahoo.wow.api.modeling.FirstOperatorCapable
 import me.ahoo.wow.api.modeling.NamedAggregate
 import me.ahoo.wow.api.modeling.OperatorCapable
 import me.ahoo.wow.api.modeling.OwnerId
+import me.ahoo.wow.api.modeling.SnapshotTimeCapable
 import me.ahoo.wow.api.modeling.StateCapable
 import me.ahoo.wow.api.modeling.TenantId
 import me.ahoo.wow.api.naming.Materialized
@@ -32,22 +34,24 @@ data class MaterializedSnapshot<S : Any>(
     override val ownerId: String = OwnerId.DEFAULT_OWNER_ID,
     val aggregateId: String,
     override val version: Int,
-    val eventId: String,
+    override val eventId: String,
     override val firstOperator: String,
     override val operator: String,
     override val firstEventTime: Long,
     override val eventTime: Long,
     override val state: S,
-    val snapshotTime: Long,
+    override val snapshotTime: Long,
     override val deleted: Boolean
 ) : NamedAggregate,
     TenantId,
     OwnerId,
     Version,
+    EventIdCapable,
     Materialized,
     FirstOperatorCapable,
     OperatorCapable,
     FirstEventTimeCapable,
     EventTimeCapable,
     StateCapable<S>,
+    SnapshotTimeCapable,
     DeletedCapable
