--- wow-api/src/test/kotlin/me/ahoo/wow/api/query/MaterializedSnapshotTest.kt
@@ -1,28 +1,0 @@
-package me.ahoo.wow.api.query
-
-import me.ahoo.test.asserts.assert
-import org.junit.jupiter.api.Test
-
-class MaterializedSnapshotTest {
-
-    @Test
-    fun test() {
-        val snapshot = MaterializedSnapshot(
-            contextName = "contextName",
-            aggregateName = "aggregateName",
-            tenantId = "tenantId",
-            ownerId = "ownerId",
-            aggregateId = "aggregateId",
-            version = 1,
-            eventId = "eventId",
-            firstOperator = "firstOperator",
-            operator = "operator",
-            firstEventTime = 1,
-            eventTime = 1,
-            state = "state",
-            snapshotTime = 1,
-            deleted = false
-        )
-        snapshot.assert().isNotNull()
-    }
-}
